requirements.txt
cat > /mnt/user-data/outputs/requirements.txt << 'EOF'
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.4
tensorflow>=2.13
tensorflow-datasets>=4.9
importlib_resources>=6.0
EOF